fix: change int to size_t for correct type comparison

+++ main.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include <stdlib.h>
 #include <stdio.h>
 #include <stdint.h>
@@ -8,19 +9,20 @@ int main(int argc, char *argv[])
     printf("%s -h for usage.\n", argv[0]);
     exit(EXIT_FAILURE);
   }
+
   FILE *f = fopen(argv[1], "rb");
   if (f == NULL) {
     perror(argv[1]);
     exit(EXIT_FAILURE);
   }
-  uint8_t buf[16]; // Declaration of array on the stack
-                   //
+
+  uint8_t buf[16];
   size_t offset = 0;
   size_t n;
   while ((n = fread(buf,1,16,f)) > 0) {
     /* HEX Printing */
     printf("%08zx ", offset);
-    for (int i = 0; i < 16; i++) {
+    for (size_t i = 0; i < 16; i++) {
       if (i < n) {
       printf("%02x ", buf[i]);
       } else{
@@ -29,7 +31,7 @@ int main(int argc, char *argv[])
     }
     /* ASCII Letter Printing */
     printf("|");
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
       if (buf[i] >= 32 && buf[i] <= 126) {
       printf("%c", buf[i]);
       } else{
